requirements.txt
anyio==3.6.2
asgiref==3.4.1
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
dnspython==2.2.1
email-validator==1.3.1
fastapi==0.83.0
greenlet==2.0.2
h11==0.13.0
httptools==0.3.0
idna==3.10
immutables==0.19
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.1
passlib==1.7.4
psycopg2==2.9.8
pydantic==1.9.2
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0.1
requests==2.27.1
six==1.17.0
sniffio==1.2.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
starlette==0.19.1
typing_extensions==4.1.1
ujson==4.0.1
urllib3==1.26.20
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0
